pyproject.toml
[tool.poetry]
name = "wavelink_discord_bot"
version = "0.3.0"
description = "Simple Discord music bot."
authors = ["I ZomBie <patrik009009@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.3.2"
wavelink = "^3.3.0"
python-dotenv = "^1.0.1"
pynacl = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
